Added packet struct definition and tidied up some function defs to be in line with Doxygen

+++ solar_Pro/common/projSensors.c
@@ -100,7 +100,7 @@ int getInternalTemperature(void){
 /*** BATTERY VOLTAGE FUNCTION INTERNAL ***/
 //function for outputting the temperature value of the zolertia remote
 //@return int: battery voltage in mV
-int getBatteryVoltage(void){
+uint16_t getBatteryVoltage(void){
 
 	return vdd3_sensor.value(CC2538_SENSORS_VALUE_TYPE_CONVERTED);
 }
